moving allowed_organizations to auth module for user
moving tests as well
