+++ src/test/models/user_test.rb
@@ -11,8 +11,6 @@
 # http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
 
 require 'test/models/user_base'
-require 'test/support/auth_support'
-
 
 class UserCreateTest < MiniTest::Rails::ActiveSupport::TestCase
   include TestUserBase
@@ -158,26 +156,4 @@ class UserDefaultEnvTest < MiniTest::Rails::ActiveSupport::TestCase
     assert User.find_by_default_environment(@env.id).empty?
     assert @user.default_environment.nil?
   end
-end
-
-
-class UserOrganizationAccess < MiniTest::Rails::ActiveSupport::TestCase
-  include AuthorizationHelperMethods
-  include TestUserBase
-
-  def setup
-    super
-    @user = @no_perms_user
-    @org = @acme_organization
-    @org2 = FactoryGirl.create(:organization, :label=>"foolabel", :name=>"fooname")
-  end
-
-  def test_access_two_orgs
-    assert @user.allowed_organizations.size == 0
-    allow(@user.own_role, [:read], :providers, nil, @org)
-    allow(@user.own_role,[:read], :providers, nil, @org2)
-    @user.save!
-    assert @user.allowed_organizations.size == 2
-  end
-
 end
